Distinguish between typed and untyped numbers
Numbers are untyped by default and only become typed when an explicit
type conversion is used (e.g. f32, u8, etc). The result of operations
will be typed if the input is typed or untyped if the input is untyped.
Binary operators (excluding shifts) will have a typed result if either
input is typed. The output width of an operation is the maximum typed
width of its inputs.